app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beanie==1.25.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
lazy-model==0.2.0
motor==3.3.2
openai==1.12.0
packaging==23.2
pluggy==1.4.0
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
pytest==8.0.0
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
